% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_functions.R
\name{data_clean_tot}
\alias{data_clean_tot}
\title{Cleans up a Sackmann point-by-point dataset}
\usage{
data_clean_tot(us_open, last_rownum, first_rownum = 1)
}
\arguments{
\item{us_open}{A Sackmann point-by-point dataset.}

\item{last_rownum}{The last row to apply the cleaning procedure.}

\item{first_rownum:}{The first row to apply the cleaning procedure. All rows
from \code{first_rownum} to \code{last_rownum} will go through the cleaning
procedure.}
}
\value{
A dataframe whose rows are the games of tennis matches with the following columns:
  \itemize{
    \item \code{pbp_id}: identification corresponding to that in the original dataset
    \item \code{match_num}: counter of the matches cleaned
    \item \code{set_num}: identifies the set in which the game is played
    \item \code{game_num}: counter of the games in each set (resets to 1 after a new set)
    \item \code{pbp}: point-by-point record of the game
    \item \code{server}: points won by the server of the game
    \item \code{returner}: points won by the returner of the game
    \item \code{winner}: winner of the game i.e. server or returner
    \item \code{player1_serve}: 1 if player1 is server, 0 otherwise
    \item \code{player1_game}: 1 if player1 wins game, 0 otherwise
    \item \code{player2_game}: 1 if player2 wins game, 0 otherwise
    \item \code{player1_game_acc_set}: running total of number of games won by player1 in a given set
    \item \code{player2_game_acc_set}: running total of number of games won by player2 in a given set
    \item \code{player1_game_acc_total}: running total of number of games won by player1 in a given match
    \item \code{player2_game_acc_total}: running total of number of games won by player2 in a given match
    \item \code{total_game_acc}: running total of number of games in a given match
    \item \code{leading_set}: denotes which player has won more games in a given set
    \item \code{player1_set_acc}: number of sets player1 wins
    \item \code{player2_set_acc}: number of sets player2 wins
  }
}
\description{
\code{data_clean_tot} converts a match-wise Sackmann point-by-point dataset
into a game-wise dataaset with point-by-point records, the outcome of the
game, and accumulation scores set-wise and match-wise.
}
\seealso{
Other sackmann data cleaning functions: \code{\link{score_sets}},
  \code{\link{split_games_by_sets}},
  \code{\link{split_sets}}
}
\concept{sackmann data cleaning functions}
