% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_functions.R
\name{get_ATP_code}
\alias{get_ATP_code}
\title{Retrieves ATP Code for each player}
\usage{
get_ATP_code(x)
}
\arguments{
\item{x}{A column of google search URLs in the form of a string for each player.}
}
\value{
ATP code for each player.
}
\description{
ATP has a player-specific code. This function retrieves that code based on a
google search of the player's name and basic search parameters to find the
ATP url where the code is located.
}
\seealso{
Other web scraping functions: \code{\link{clean_data}},
  \code{\link{extract_data}}, \code{\link{get_ATP_url}},
  \code{\link{get_plus_name}},
  \code{\link{manipulate_data}}
}
\concept{web scraping functions}
