% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_functions.R
\name{get_ATP_url}
\alias{get_ATP_url}
\title{Create ATP URL from ATP Code}
\usage{
get_ATP_url(x)
}
\arguments{
\item{x}{ATP code as a string}
}
\value{
The URL of the player stats on the ATP website
}
\description{
The function generates the URL that links a player's name to his player stats profile on the ATP website.
}
\seealso{
Other web scraping functions: \code{\link{clean_data}},
  \code{\link{extract_data}}, \code{\link{get_ATP_code}},
  \code{\link{get_plus_name}},
  \code{\link{manipulate_data}}
}
\concept{web scraping functions}
