% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_functions.R
\name{get_plus_name}
\alias{get_plus_name}
\title{Replaces spaces in names with "+"}
\usage{
get_plus_name(x)
}
\arguments{
\item{x}{A column of names as strings (note: not factors).}
}
\value{
Full name separated by "+" instead of " ".
}
\description{
\code{get_plus_name} takes a full name in the form of a string and converts
it into a string that has "+" instead of spaces separating the name. The
output is meant to be pasted into a google search URL.
}
\seealso{
Other web scraping functions: \code{\link{clean_data}},
  \code{\link{extract_data}}, \code{\link{get_ATP_code}},
  \code{\link{get_ATP_url}}, \code{\link{manipulate_data}}
}
\concept{web scraping functions}
