% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_functions.R
\name{manipulate_data}
\alias{manipulate_data}
\title{Creates a dataframe of ATP urls and ATP codes given the names of players}
\usage{
manipulate_data(x)
}
\arguments{
\item{x}{Dataframe of players we are interested in.}
}
\value{
Dataframe with the \code{google_url}, \code{get_ATP_code}, and
  \code{get_ATP_url} for every player
}
\description{
Given the names of the players, this function creates the Google URLS,
identifies the ATP codes, and generates the ATP URLs.
}
\seealso{
Other web scraping functions: \code{\link{clean_data}},
  \code{\link{extract_data}}, \code{\link{get_ATP_code}},
  \code{\link{get_ATP_url}}, \code{\link{get_plus_name}}
}
\concept{web scraping functions}
