% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbp_raw.df}
\alias{pbp_raw.df}
\title{Point-by-point records of tennis matches}
\format{A data frame with 13050 rows and 12 variables:
\describe{
  \item{pbp_id}{unique identifier of a tennis match}
  \item{date}{date of match, format: DD MON YY}
  \item{tny_name}{tournament name}
  \item{tour}{name of tour, i.e., ATP}
  \item{draw}{draw, i.e., Main}
  \item{server1}{name of player who served first, often referred to as player1}
  \item{server2}{name of opponent of \code{server1}, often referred to as player2}
  \item{winner}{winner of match, i.e., 1 or 2, corresponding to one of the previous two columns}
  \item{pbp}{point-by-point record; S (server won), R (returner won), A (ace), D (double fault); Sets are delimited with the '.' character, games are delimited with the ';' character, and the '/' character indicates changes of serve during a tiebreak}
  \item{score}{final score of match}
  \item{adf_flag}{1 if the point sequence notes any aces or double faults, 0 if not}
  \item{wh_minutes}{...}
}}
\source{
\url{https://github.com/JeffSackmann/tennis_pointbypoint}
}
\usage{
pbp_raw.df
}
\description{
A dataset made from combining `pbp_matches_atp_main_current.csv` and
`pbp_matches_atp_main_archive` by Jeff Sackmann. Contains point-by-point
records of ATP tennis matches from 2011 to 2017.
}
\seealso{
\code{\link{pbp_raw_clean.df}} for cleaned version
}
\concept{Sackmann original files}
\keyword{datasets}
