% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbp_raw_clean.df}
\alias{pbp_raw_clean.df}
\title{Cleaned version of Sackmann's \code{pbp_raw.df}}
\format{A data frame with 364797 rows and 19 variables:
\describe{
    \item{pbp_id}{identification corresponding to that in the original dataset}
    \item{match_num}{counter of the matches cleaned}
    \item{set_num}{identifies the set in which the game is played}
    \item{game_num}{counter of the games in each set (resets to 1 after a new set)}
    \item{pbp}{point-by-point record of the game; S (server won), R (returner won), A (ace), D (double fault)}
    \item{server}{points won by the server of the game}
    \item{returner}{points won by the returner of the game}
    \item{winner}{winner of the game i.e. server or returner}
    \item{player1_serve}{1 if player1 is server, 0 otherwise}
    \item{player1_game}{1 if player1 wins game, 0 otherwise}
    \item{player2_game}{1 if player2 wins game, 0 otherwise}
    \item{player1_game_acc_set}{running total of number of games won by player1 in a given set}
    \item{player2_game_acc_set}{running total of number of games won by player2 in a given set}
    \item{player1_game_acc_total}{running total of number of games won by player1 in a given match}
    \item{player2_game_acc_total}{running total of number of games won by player2 in a given match}
    \item{total_game_acc}{running total of number of games in a given match}
    \item{leading_set}{denotes which player has won more games in a given set}
    \item{player1_set_acc}{number of sets player1 wins}
    \item{player2_set_acc}{number of sets player2 wins}
}}
\usage{
pbp_raw_clean.df
}
\description{
A game-wise manipulation of Sackmann's pbp_raw.df
}
\seealso{
\code{\link{pbp_raw.df}} for original file, \code{\link{data_clean_tot}} for cleaning function
}
\concept{Sackmann manipulated files}
\keyword{datasets}
