% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{players}
\alias{players}
\title{Match-wise pairing of \code{server1} and \code{server2}}
\format{A data frame with 7194 rows and 3 variables:
\describe{
  \item{pbp_id4}{pbp_id for matches in which there is a set-wise difference of at least 4 games between the lagging and leading player}
  \item{server1}
  \item{server2}
}}
\usage{
players
}
\description{
A data frame that shows who the players are for every match that has a set-wise game deficit of at least 4
}
\concept{Sackmann manipulated files, player stats}
\keyword{datasets}
