% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{players_stats}
\alias{players_stats}
\title{Players' Statistics}
\format{A data frame of 887 rows and 27 variables
\describe{
  \item{name}{First and Last Name}
  \item{rank}{Rank}
  \item{age}{Age and Birth Date}
  \item{pro_start}{Pro_start}
  \item{weight}{Weight}
  \item{height}{Height}
  \item{residence}{Residence}
  \item{hand}{Hand}
  \item{coach}{Coach}
  \item{aces}{Aces}
  \item{df}{Double Faults}
  \item{first_serve}{1st Serve}
  \item{first_serve_won}{1st Serve Points Won}
  \item{second_serve_won}{2nd Serve Points Won}
  \item{bp_faced}{Break Points Faced}
  \item{bp_saved}{Break Points Saved}
  \item{serv_game_played}{Service Games Played}
  \item{serv_game_won}{Service Games Won}
  \item{total_serv_won}{Total Service Points Won}
  \item{first_return}{1st Serve Return Points Won}
  \item{second_return}{2nd Serve Return Points Won}
  \item{bp_opp}{Break Points Opportunities}
  \item{bp_conv}{Break Points Converted}
  \item{ret_game_played}{Return Games Played}
  \item{ret_game_won}{Return Games Won}
  \item{ret_won}{Return Points Won}
  \item{total_ret_won}{Total Points Won}
}}
\source{
\url{http://atpworldtour.com}
}
\usage{
players_stats
}
\description{
Web-scraped statistics from ATP website re: players in matches that have a set-wise game deficit of at least 4
}
\concept{player stats}
\keyword{datasets}
