% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_functions.R
\name{score_sets}
\alias{score_sets}
\title{Identifies the winner of each game}
\usage{
score_sets(setnum)
}
\arguments{
\item{setnum}{A data frame whose rows contain pbp records of a game.}
}
\value{
The original data frame with an extra column `winner` denoting
  whether the server or returner won the game; NAs exist in `winner` if set4
  and/or set5 have 99 in `pbp`.
}
\description{
For each game, identify whether the server or returner won by counting who
scored more points than the other.
}
\seealso{
Other sackmann data cleaning functions: \code{\link{data_clean_tot}},
  \code{\link{split_games_by_sets}},
  \code{\link{split_sets}}
}
\concept{sackmann data cleaning functions}
