% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_functions.R
\name{split_games_by_sets}
\alias{split_games_by_sets}
\title{Create 5 variables for each set}
\usage{
split_games_by_sets(vec)
}
\arguments{
\item{vec}{A list of strings, each representing the pbp record in each set.}
}
\value{
5 variables called `set1`, `set2`, `set3`, `set4`, and `set5`. Each
  are pbp records of the games in the respective set.
}
\description{
For each match, \code{split_games_by_sets} creates 5 variables that store the
  pbp records of each set in the match; If set4 and set5 are not realized, then
  let pbp be `99`.
}
\seealso{
Other sackmann data cleaning functions: \code{\link{data_clean_tot}},
  \code{\link{score_sets}}, \code{\link{split_sets}}
}
\concept{sackmann data cleaning functions}
