% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_functions.R
\name{split_sets}
\alias{split_sets}
\title{Separates a pbp (point-by-point) string into sets}
\usage{
split_sets(vec.pbp)
}
\arguments{
\item{vec.pbp}{A vector of strings containing the pbp record of tennis matches.}
}
\value{
A list of at most 5 strings, each representing the pbp record of sets
  in a tennis match.
}
\description{
\code{split_sets} separates a string containing the pbp record of tennis
matches whose sets are delimited by a period ("."). Since there is a maximum
of 5 sets tennis match, expect a list of 5 entries, each with their pbp
record.
}
\seealso{
Other sackmann data cleaning functions: \code{\link{data_clean_tot}},
  \code{\link{score_sets}},
  \code{\link{split_games_by_sets}}
}
\concept{sackmann data cleaning functions}
