% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis}
\alias{tennis}
\title{Cleaned game-wise dataset with player statistics}
\format{A data frame with 17680 rows and 81 variables. Variables that are character specific have been omitted (e.g., ace_lagging and ace_leading have the same description, but refer to different players hence only ace_lagging appears in the list below):
\describe{
  \item{pbp_id}{pbp_id for matches in which there is a set-wise difference of at least 4 games between the lagging and leading player}
  \item{match_num}{counter of the matches cleaned from \code{pbp_raw_clean.df}}
  \item{set_num}{identifies the set in which the game is played}
  \item{game_num}{counter of the games in each set (resets to 1 after a new set)}
  \item{pbp}{point-by-point record of the game; S (server won), R (returner won), A (ace), D (double fault)}
  \item{server}{points won by the server of the game}
  \item{returner}{points won by the returner of the game}
  \item{winner}{winner of the game i.e. server or returner}
  \item{player1_serve}{1 if player1 is server, 0 otherwise}
  \item{player1_game}{1 if player1 wins game, 0 otherwise}
  \item{player2_game}{1 if player2 wins game, 0 otherwise}
  \item{player1_game_acc_set}{running total of number of games won by player1 in a given set}
  \item{player2_game_acc_set}{running total of number of games won by player2 in a given set}
  \item{player1_game_acc_total}{running total of number of games won by player1 in a given match}
  \item{player2_game_acc_total}{running total of number of games won by player2 in a given match}
  \item{total_game_acc}{running total of number of games in a given match}
  \item{leading_set}{denotes which player has won more games in a given set}
  \item{player1_set_acc}{number of sets player1 wins}
  \item{player2_set_acc}{number of sets player2 wins}
  \item{diff_game_set}{set-wise game deficit; by construction, must be of magnitude at least four}
  \item{match_winner}{winner of the match, i.e., 1 or 2}
  \item{player1_leading_set}{1 if player1 has more games than player2 in a given set}
  \item{rally_length}{length of rally, i.e., nchar(pbp)}
  \item{server1}{Name of server 1}
  \item{server2}{Name of server 2}
  \item{rank_lagging}{Rank}
  \item{age_lagging}{Age and Birth Date}
  \item{pro_start_lagging}{Pro_start}
  \item{weight_lagging}{Weight}
  \item{height_lagging}{Height}
  \item{residence_lagging}{Residence}
  \item{hand_lagging}{Hand}
  \item{coach_lagging}{Coach}
  \item{aces_lagging}{Aces}
  \item{df_lagging}{Double Faults}
  \item{first_serve_lagging}{1st Serve}
  \item{first_serve_won_lagging}{1st Serve Points Won}
  \item{second_serve_won_lagging}{2nd Serve Points Won}
  \item{bp_faced_lagging}{Break Points Faced}
  \item{bp_saved_lagging}{Break Points Saved}
  \item{serv_game_played_lagging}{Service Games Played}
  \item{serv_game_won_lagging}{Service Games Won}
  \item{total_serv_won_lagging}{Total Service Points Won}
  \item{first_return_lagging}{1st Serve Return Points Won}
  \item{second_return_lagging}{2nd Serve Return Points Won}
  \item{bp_opp_lagging}{Break Points Opportunities}
  \item{bp_conv_lagging}{Break Points Converted}
  \item{ret_game_played_lagging}{Return Games Played}
  \item{ret_game_won_lagging}{Return Games Won}
  \item{ret_won_lagging}{Return Points Won}
  \item{total_ret_won_lagging}{Total Points Won}
}}
\usage{
tennis
}
\description{
A combination of player statistics and game-wise pbp data in which there is a lagging and a leading player
}
\seealso{
\code{\link{tennis_cleaned}} for a player1/player2-agnostic dataset

Other final files: \code{\link{tennis_cleaned}},
  \code{\link{tennis_data}}
}
\concept{final files}
\keyword{datasets}
