% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tennis_cleaned}
\alias{tennis_cleaned}
\title{Cleaned version of \code{tennis}}
\format{A data frame with 17680 rows and 107 variables; only variables not in \code{tennis} are explained below. Note that differences were calculated by subtracting leading numbers from lagging numbers.
\describe{
  \item{see \code{tennis}}{}
  \item{lagging_serve}{1 if lagging player serves, 0 otherwise}
  \item{leading_serve}{1 if leading player serves, 0 otherwise}
  \item{lagging_points}{points scored by lagging player in the game}
  \item{leading_points}{points scored by leading player in the game}
  \item{lagging_game}{1 if lagging player wins game, 0 otherwise}
  \item{leading_game}{1 if leading player wins game, 0 otherwise}
  \item{lagging_game_acc_set}{number of games won by the lagging player in the given set thus far}
  \item{leading_game_acc_set}{number of games won by the leading player in the given set thus far}
  \item{lagging_game_acc_tot}{number of games won by the lagging player in entire match thus far}
  \item{leading_game_acc_tot}{number of games won by the leading player in entire match thus far}
  \item{lagging_set}{1 if lagging player wins set, 0 otherwise}
  \item{leading_set}{1 if leading player wins set, 0 otherwise}
  \item{lagging_set_acc}{number of sets won by the lagging player thus far}
  \item{leading_set_acc}{number of sets won by the leading player thus far}
  \item{lagging_match_win}{1 if lagging player wins match}
  \item{leading_match_win}{1 if leading player wins match}
  \item{diff_game_tot}{match-wise game deficit}
  \item{diff_set}{match-wise set deficit}
  \item{lagging_hand}{1 if lagging player is right-hand dominant, 0 otherwise}
  \item{leading_hand}{1 if leading player is right-hand dominant, 0 otherwise}
  \item{diff_hand}{1 if players are dominant in different hands, 0 otherwise}
  \item{diff_rank}{difference in the players' ranks}
  \item{diff_serve_won}{difference in the service games won}
  \item{diff_ret_won}{difference in return games won}
  \item{lagging_age}{age of lagging player}
  \item{leading_age}{age of leading player}
  \item{diff_age}{difference in the players' ages}
}}
\usage{
tennis_cleaned
}
\description{
This dataset has lagging-leading differences in player statistics and renames descriptions of the game in terms of lagging and leading player as opposed to player1 and player2.
}
\seealso{
\code{\link{tennis_cleaned}} for a player1/player2-defined dataset; \code{\link{tennis_data}} for child set

Other final files: \code{\link{tennis_data}},
  \code{\link{tennis}}
}
\concept{final files}
\keyword{datasets}
